more code hygiene, clean up redundant or unnecessary code

+++ commons/marmotta-sesame-tools/marmotta-rio-rss/src/ext/java/com/sun/syndication/io/impl/RSS094Parser.java
@@ -57,7 +57,7 @@ public class RSS094Parser extends RSS093Parser {
                 //let it go by
             }
             if (ttlValue != null) {
-                channel.setTtl(ttlValue.intValue());
+                channel.setTtl(ttlValue);
             }
         }
